fix(web): route waitlist storage through platform adapter
